Fix issue #5 and fix failing tests.
Issue #5
--------
Do not inherit indentation of the previous line if the current line
has less indentation but is valid.

Failing tests
-------------
Commented out in commit #6e60ef0. The tests were failing because
the indent function
    GetPythonPEPIndent()
can modify the cursor position. There is nothing wrong with that
consequence, vim takes that into account and restores the cursor
position. Therefore when we call the function we need to restore
the cursor position.