Fix failing commented out tests
Related commits
---------------
- #6e60ef0

Reason
------
The indent functions, in this case
      GetPythonPEPIndent()
can modify the cursor position. There is nothing wrong with that
consequence, vim takes it into account and restores the cursor
position (see References below). For the tests vim is not doing
the calling, so the cursor is not restored.

Solution
--------
Restore the cursor position after calling the indent function.