Fix issue #5 (Correct indent after end of block)
Problem
-------
Given two adjacent blocks where the 2nd block has
an equal or smaller indentation level compared to the 1st
block. If the blocks are not separated by an empty line,
the 2nd block will be placed (indented) inside the 1st
block.

This changes the meaning of code that already adheres to the
pep8 standard.

Solution
--------
Do not inherit indentation of the previous line if the current line
has less indentation but is valid.