Add option for hanging closing brackets (#94)
The main idea is discussed at length in PyCQA/pycodestyle#103.

* Add tests for `python_pep8_indent_hang_closing`

This might be overkill. It reruns the current "vim" set of tests once
for each value of the new setting. Of course, this makes the test suite
take longer to run. I couldn't think of a good way to factor out the
relevant test cases without creating a mess.